rename box2d file to box
