+++ packages/editor/src/index.ts
@@ -267,7 +267,7 @@ export {
 	type SelectionCorner,
 	type SelectionEdge,
 	type SelectionHandle,
-} from './lib/primitives/Box2d'
+} from './lib/primitives/Box'
 export { Matrix2d, type Matrix2dModel } from './lib/primitives/Matrix2d'
 export { Vec, type VecLike } from './lib/primitives/Vec'
 export { EASINGS } from './lib/primitives/easings'
