+++ packages/editor/src/lib/components/default-components/DefaultSelectionForeground.tsx
@@ -3,7 +3,7 @@ import classNames from 'classnames'
 import { ComponentType, useRef } from 'react'
 import { useEditor } from '../../hooks/useEditor'
 import { useTransform } from '../../hooks/useTransform'
-import { Box2d } from '../../primitives/Box2d'
+import { Box2d } from '../../primitives/Box'
 import { toDomPrecision } from '../../primitives/utils'
 
 /** @public */
