+++ packages/editor/src/lib/editor/shapes/shared/arrow/curved-arrow.ts
@@ -1,5 +1,5 @@
 import { TLArrowShape } from '@tldraw/tlschema'
-import { Box2d } from '../../../../primitives/Box2d'
+import { Box2d } from '../../../../primitives/Box'
 import { Matrix2d } from '../../../../primitives/Matrix2d'
 import { Vec, VecLike } from '../../../../primitives/Vec'
 import { intersectCirclePolygon, intersectCirclePolyline } from '../../../../primitives/intersect'
