+++ packages/editor/src/lib/primitives/Box.test.ts
@@ -1,4 +1,4 @@
-import { Box2d } from './Box2d'
+import { Box2d } from './Box'
 import { Vec } from './Vec'
 
 describe('Box2d', () => {
