+++ packages/editor/src/lib/primitives/geometry/Circle2d.ts
@@ -1,4 +1,4 @@
-import { Box2d } from '../Box2d'
+import { Box2d } from '../Box'
 import { Vec } from '../Vec'
 import { intersectLineSegmentCircle } from '../intersect'
 import { PI2 } from '../utils'
