+++ packages/editor/src/lib/primitives/geometry/Geometry2d.ts
@@ -1,4 +1,4 @@
-import { Box2d } from '../Box2d'
+import { Box2d } from '../Box'
 import { Vec } from '../Vec'
 import { pointInPolygon } from '../utils'
