+++ packages/editor/src/lib/primitives/utils.ts
@@ -1,4 +1,4 @@
-import { Box2d } from './Box2d'
+import { Box2d } from './Box'
 import { Vec, VecLike } from './Vec'
 
 /** @public */
